requirements.txt


# Do """sudo apt install portaudio19-dev python3-pyaudio && sudo apt install ffmpeg libavcodec-extra"""
# Do """sudo apt install git"""
# sudo raspi-config nonint do_spi 0


# Change default sound card to https://superuser.com/a/630048  --- doesn't work
pydub==0.25.1
#pyaudio

# Jukebox Core
# For USB inputs (reader, buttons) and bluetooth buttons
#evdev
pyalsaaudio
#pulsectl
#python_mpd2
#ruamel.yaml
# For playlistgenerator
#requests
# For music_cover_art
#eyed3
# For the publisher event reactor loop:
#tornado

# RPi's GPIO packages: these are installed via APT on the PI
# On regular machines, install them manually if needed for development
RPi.GPIO
# gpiozero

# --------
# RC522 related requirements
# You need to install these with `sudo python3 -m pip install --upgrade --force-reinstall -q -r requirements.txt`

# pi-rc522 use latest version from Github
# This is the original versions. Seems unmaintained for the moment
# git+https://github.com/ondryaso/pi-rc522.git#egg=pi-rc522

# The fork of kevinvalk has some good improvements
# https://github.com/kevinvalk/pi-rc522
# Get the kevinvalk fork yet again from a different fork which ensures compatibility with the Phoniebox
git+https://github.com/ChisSoc/pi-rc522.git#egg=pi-rc522
